requirements.txt
kubernetes
requests
PyYAML